Update poetry lock and move models into other dier
