+++ bunq_ynab_connect/data/bunq_account_to_ynab_account_mapper.py
@@ -8,8 +8,8 @@ from bunq.sdk.model.generated.endpoint import (
 from kink import inject
 
 from bunq_ynab_connect.data.storage.abstract_storage import AbstractStorage
-from bunq_ynab_connect.models.ynab.bunq_account import BunqAccount
-from bunq_ynab_connect.models.ynab.ynab_account import YnabAccount
+from bunq_ynab_connect.models.bunq_account import BunqAccount
+from bunq_ynab_connect.models.ynab_account import YnabAccount
 
 
 @inject
