+++ bunq_ynab_connect/data/data_extractors/ynab_budget_extractor.py
@@ -8,7 +8,7 @@ from ynab import ApiClient
 from bunq_ynab_connect.clients.ynab_client import YnabClient
 from bunq_ynab_connect.data.data_extractors.abstract_extractor import AbstractExtractor
 from bunq_ynab_connect.data.storage.abstract_storage import AbstractStorage
-from bunq_ynab_connect.models.ynab.ynab_budget import YnabBudget
+from bunq_ynab_connect.models.ynab_budget import YnabBudget
 
 
 class YnabBudgetExtractor(AbstractExtractor):
