+++ bunq_ynab_connect/models/matched_transaction.py
@@ -2,8 +2,8 @@ from pydantic import BaseModel, validator
 from pydantic.dataclasses import dataclass
 
 from bunq_ynab_connect.helpers.general import date_to_datetime
-from bunq_ynab_connect.models.ynab.bunq_payment import BunqPayment
-from bunq_ynab_connect.models.ynab.ynab_transaction import YnabTransaction
+from bunq_ynab_connect.models.bunq_payment import BunqPayment
+from bunq_ynab_connect.models.ynab_transaction import YnabTransaction
 
 
 class MatchedTransaction(BaseModel):
