+++ bunq_ynab_connect/models/ynab_transaction.py
