+++ bunq_ynab_connect/sync_bunq_to_ynab/payment_syncer.py
@@ -17,8 +17,8 @@ from bunq_ynab_connect.data.bunq_account_to_ynab_account_mapper import (
     BunqAccountToYnabAccountMapper,
 )
 from bunq_ynab_connect.data.storage.abstract_storage import AbstractStorage
-from bunq_ynab_connect.models.ynab.bunq_payment import BunqPayment
-from bunq_ynab_connect.models.ynab.ynab_account import YnabAccount
+from bunq_ynab_connect.models.bunq_payment import BunqPayment
+from bunq_ynab_connect.models.ynab_account import YnabAccount
 from bunq_ynab_connect.sync_bunq_to_ynab.payment_queue import PaymentQueue
